Add MemoryBtn file.
